Fix GetLastSeenString method (#103)
Using a double for calculating if there are any weeks because if there is not a full extra day it will not display the weeks.

+++ Jellyfin.Plugin.PlaybackReporting/Api/PlaybackReportingActivityController.cs
@@ -571,7 +571,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Api
                 last_seen += GetTimePart((int)(span.TotalDays / 365), "year");
             }
 
-            if ((int)(span.TotalDays % 365) > 7)
+            if ((double)(span.TotalDays % 365) > 7)
             {
                 last_seen += GetTimePart((int)((span.TotalDays % 365) / 7), "week");
             }
